Correction de typo, retrait d'un debug
Former-commit-id: cd97a69a177e3d4fb2f73aaaaf60cff8338dab3b

+++ modelConvert/SKModelCommons.cpp
@@ -34,7 +34,7 @@ double SKModel<EQT>::rk4(double step, double xCurr, double xPrev, double yPrev) 
 
 template<class EQT>
 int SKModel<EQT>::findLatencyIndex(double latency) {
-	float currTime = Marking.P->_TIME[Marking.P->lastEntry];
+	double currTime = Marking.P->_TIME[Marking.P->lastEntry];
 	while (Marking.P->countDown > 0 && Marking.P->_TIME[Marking.P->countDown] > (currTime - latency)) {
 		Marking.P->countDown--;
 	}
